WiFi: Side effect fixed from the "Fixed wrong "WiFi Direct" state..."
This patch fixes side effect which was introduced by
4541d9053bb85190dbd845aebba988eeb64b0ba8 commit ID.

Side effect description:
If Wi-Fi is ON and we try to enable Wi-Fi Direct,
WiFi should be turned Off only if WiFi Direct selection confirmed
by user. But WiFi is always turned Off despite the user selection.
Side effect has took place because of wrong placement of the fix
mentioned above.

Change-Id: I4b996c3c8fc09498100657cb5dd8380a5944f6df
Signed-off-by: Anatolii.Shuba <x0158321@ti.com>